07-11-2022

+++ app/Http/Controllers/Api/FlashSaleController.php
@@ -13,23 +13,27 @@ use Carbon\Carbon;
 class FlashSaleController extends Controller
 {
     public function index(){
-        
         $flashsale_all = FlashSaleProduct::join('flashsale_frame', 'flashsale_frame.id_flashsale_frame', '=','flashsale_product.id_flashsale_frame')->get();      
         $now = Carbon::now();
-        foreach ($flashsale_all as $key => $flashsale) {
-            $dateStart = Carbon::createFromFormat('Y-m-d H:i:s', $flashsale->sale_day.' '.$flashsale->start,'Asia/Ho_Chi_Minh');
-            $dateEnd = Carbon::createFromFormat('Y-m-d H:i:s', $flashsale->sale_day.' '.$flashsale->end,'Asia/Ho_Chi_Minh');
-            if($dateEnd->lt($now) == true){
-                $product_sale = Product::where('id_product', '=', $flashsale->id_product)->first();
-                $product_sale->update(['discount' => 0.0]);
-                $flashsale->delete();
-            }
-            if($dateStart->gte($now) == true){
-                $product_sale = Product::where('id_product', '=', $flashsale->id_product)->first();
-                $product_sale->update(['discount' => $flashsale->salePercent]);
+        if($flashsale_all){
+            foreach ($flashsale_all as $key => $flashsale) {
+                $dateStart = Carbon::createFromFormat('Y-m-d H:i:s', $flashsale->sale_day.' '.$flashsale->start,'Asia/Ho_Chi_Minh');
+                $dateEnd = Carbon::createFromFormat('Y-m-d H:i:s', $flashsale->sale_day.' '.$flashsale->end,'Asia/Ho_Chi_Minh');
+                if($dateEnd->lt($now) == true){
+                    $product_sale = Product::where('id_product', '=', $flashsale->id_product)->first();
+                    if($product_sale){
+                        $product_sale->update(['discount' => 0.0]);
+                    }
+                    $flashsale->delete();
+                }
+                if($now->gte($dateStart)){
+                    $product_sale = Product::where('id_product', '=', $flashsale->id_product)->first();
+                    if($product_sale){
+                        $product_sale->update(['discount' => $flashsale->salePercent]);
+                    }
+                }
             }
         }
-
         $flashsale_product = FlashSaleProduct::join('flashsale_frame', 'flashsale_frame.id_flashsale_frame', '=','flashsale_product.id_flashsale_frame')
         ->join('products', 'products.id_product' , '=' , 'flashsale_product.id_product')->get();
